pyproject.toml
[project]
name = "agent-worker"
version = "1.0.0"
description = "Agentic AI with LangGraph, MCP Tools, and Memory Systems"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "mcp>=1.0.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agent*", "mcp_server*"]
